refactor(#512): extract scanner configuration to scanner_config.py
- Create scanner_config.py with:
  - Config loaders (load_scanner_config, load_trading_modes, etc.)
  - Dataclasses (TierLimits, TieredWatchlistConfig, ScanConfig, ScanResult)
  - Config path constants
- Update scanner_agent.py to import from new module
- Reduce scanner_agent.py from 969 to 813 lines (16% reduction)
