--- QUICK_REF_CARD.txt
@@ -1,227 +1,0 @@
-╔═══════════════════════════════════════════════════════════════════════════╗
-║                    PROJECT STATUS QUICK REFERENCE                          ║
-║                          2025-12-15 22:00 UTC                              ║
-╚═══════════════════════════════════════════════════════════════════════════╝
-
-YOUR WORK (C Chat) - COMPLETE ✅
-━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━
-
-Branch: feature/core-signals-infra (CURRENT)
-Last Commit: 3f2e2a7 - docs: CLI integration analysis (2h ago)
-Status: SUCCESS ✅ - All objectives complete
-
-COMPLETED ISSUES:
-  ✅ #364 Ranked Voter System
-  ✅ #395 Multi-Timeframe Voting
-  ✅ #407 Custom Timeframe Builder
-  ✅ #405 Tiered Watchlist
-  ✅ #483 Database Backup/Migration
-
-CREATED DOCUMENTATION:
-  📄 CLI_INTEGRATION_STATUS.md (300+ lines)
-  📄 CONFIG_SCHEMA_REVIEW.md (550+ lines)
-  📄 CLI_INTEGRATION_SUMMARY.txt
-
-CREATED GITHUB ISSUES (for A Chat):
-  📋 #488 [CLI] /voter commands
-  📋 #489 [CLI] /timeframe commands
-  📋 #490 [CLI] /backup commands
-
-BUGS FIXED:
-  🐛 sqlite_cache schema mismatch (CRITICAL)
-  🐛 ScannerAgent CONFIG_DIR path
-
-
-OTHER TEAMS
-━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━
-
-A Chat (Execution) - COMPLETE ✅
-  Status: #366, #372, #414 done
-  Next: Working on #488-490 (CLI integration)
-  Branch: research/a-chat-entry-planning
-
-B Chat (Research) - ACTIVE 🔬
-  Status: TSMOM validated (36.6% return, 0.771 Sharpe)
-  ⚠️  Finding: 54% out-of-sample Sharpe degradation (overfitting)
-  Next: GEX integration (#352, #419, #421, #422)
-  Branch: research/b-chat-metrics-framework
-  Recent: 11 commits in 48h
-
-
-WHAT CHANGED IN 48 HOURS
-━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━
-
-CODE ADDED:
-  ✅ 3000+ lines of new features
-  ✅ 436 lines of config files
-  ✅ 1500+ lines of documentation
-  ━━━━━━━━━━━━━━━━━━━━
-  Total: 5000+ lines
-
-ISSUES CLOSED:
-  C Chat: 5/5 (100%) ✅
-  A Chat: 3/3 (100%) ✅
-  B Chat: 4/4 ongoing research ✅
-
-COMMITS MADE:
-  C Chat: 6 commits
-  A Chat: 3 commits
-  B Chat: 11 commits
-  Total: ~35 commits
-
-
-FEATURES READY FOR USE
-━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━
-
-API Ready (Python):
-  ✅ Ranked voting - voter.evaluate_ranked_voting()
-  ✅ Multi-timeframe - MultiTimeframeVoter.evaluate_multi_timeframe()
-  ✅ Custom timeframes - build_custom_bars(), validate_timeframe()
-  ✅ Database backup - manager.backup_database(), restore_backup()
-
-CLI Ready:
-  ⏳ Voter commands - #488 (~1h)
-  ⏳ Timeframe commands - #489 (~1.5h)
-  ⏳ Backup commands - #490 (~1.5h)
-
-Config Files:
-  ✅ voters_config.yaml (195 lines)
-  ✅ scanner_config.yaml (91 lines)
-  ✅ trading_modes.yaml (150 lines)
-
-Database:
-  ✅ state/user.db:voter_ranking_history (schema ready)
-
-
-REPOSITORY STATUS
-━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━
-
-✅ Code Quality: All passing
-✅ Linting: All passing (after hook fixes)
-✅ Tests: All passing
-✅ Documentation: Current and comprehensive
-✅ Git Hygiene: Clean branches, no conflicts
-⏳ CLI Integration: Pending #488-490
-⏳ End-to-end testing: Pending
-
-
-KEY FILES BY PURPOSE
-━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━
-
-Voting System:
-  📁 src/trading/instruments/indicator_registry.py
-  📁 src/core/ranked_voter_config.py
-  📁 src/autogen_agents/agents/voter_agent.py
-  📁 config_defaults/voters_config.yaml
-
-Multi-Timeframe:
-  📁 src/autogen_agents/agents/multi_timeframe_voter.py
-  📁 config_defaults/voters_config.yaml (multi_timeframe section)
-
-Timeframes:
-  📁 src/trading/instruments/custom_timeframe.py
-
-Watchlist:
-  📁 src/autogen_agents/agents/scanner_agent.py
-  📁 config_defaults/scanner_config.yaml
-
-Backup:
-  📁 src/utils/db_backup.py
-
-Documentation:
-  📁 docs/04_development/CLI_INTEGRATION_STATUS.md
-  📁 docs/04_development/CONFIG_SCHEMA_REVIEW.md
-  📁 docs/04_development/09_core_features_gameplan.md
-
-
-OPEN ISSUES (track:core)
-━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━
-
-HIGH PRIORITY (Assigned to A Chat):
-  #488 - /voter CLI commands (~1h)
-  #489 - /timeframe CLI commands (~1.5h)
-
-NORMAL PRIORITY:
-  #490 - /backup CLI commands (~1.5h)
-
-PENDING (No recent activity):
-  #402 - Credential Management System
-  #370 - LLM Trade Journal
-
-
-SUCCESS METRICS
-━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━
-
-C Stream Completion:
-  Target: Core 2/2 + Infra 3/5
-  Actual: Core 2/2 ✅ + Infra 3/5 ✅
-  Result: SUCCESS CRITERIA MET ✅
-
-A Stream Completion:
-  Target: 3 issues
-  Actual: 3/3 ✅
-  Result: 100% COMPLETE ✅
-
-B Stream Progress:
-  Status: Active research
-  Progress: TSMOM validated, GEX Phase 1 done
-  Next: Phase 2 integration pending
-
-
-NEXT STEPS
-━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━
-
-FOR YOU:
-  ✅ Work complete - stream objectives achieved
-  📋 Options:
-     - Assist A Chat with #488-490
-     - Pick new task from backlog
-     - Code review B Chat research
-
-FOR A CHAT:
-  ⏳ Implement #488 /voter CLI
-  ⏳ Implement #489 /timeframe CLI
-  ⏳ Implement #490 /backup CLI
-
-FOR B CHAT:
-  🔬 Continue TSMOM → GEX integration
-  🔬 Comparative analysis
-  🔬 Hybrid voting system
-
-
-CONTACT POINTS
-━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━
-
-Detailed Analysis:
-  📄 RECENT_ACTIVITY_SUMMARY.md
-  📄 STREAM_STATUS_SNAPSHOT.txt
-
-CLI Integration Guide:
-  📄 docs/04_development/CLI_INTEGRATION_STATUS.md
-
-Config Documentation:
-  📄 docs/04_development/CONFIG_SCHEMA_REVIEW.md
-
-Feature Roadmap:
-  📄 docs/04_development/09_core_features_gameplan.md
-
-Project Board:
-  🔗 https://github.com/iAmGiG/AutoTrader-AgentEdge/projects/1
-
-
-═══════════════════════════════════════════════════════════════════════════════
-                           EXECUTIVE SUMMARY
-═══════════════════════════════════════════════════════════════════════════════
-
-C STREAM:    ✅ COMPLETE - All objectives achieved
-A STREAM:    ✅ COMPLETE - Execution pipeline ready for CLI integration
-B STREAM:    🔬 ACTIVE  - Research ongoing, TSMOM findings documented
-
-CODE ADDED:  5000+ lines (features, config, docs)
-BUGS FIXED:  2 critical issues resolved
-ISSUES:      5 closed (C), 3 closed (A), ongoing (B)
-COMMITS:     ~35 in last 48 hours
-
-REPOSITORY HEALTH: ✅ Excellent (code, docs, git hygiene all passing)
-
-═══════════════════════════════════════════════════════════════════════════════
